src/test/java/co/com/ies/pagination/infrastructure/primary/RestMicrolitePageTest.java
package co.com.ies.pagination.infrastructure.primary;

import static org.assertj.core.api.Assertions.*;
import static co.com.ies.pagination.domain.MicrolitePagesFixture.*;

import java.util.function.Function;

import org.junit.jupiter.api.Test;

import co.com.ies.UnitTest;
import co.com.ies.JsonHelper;
import co.com.ies.error.domain.MissingMandatoryValueException;

@UnitTest
class RestMicrolitePageTest {

  @Test
  void shouldNotConvertWithoutSourcePage() {
    assertThatThrownBy(() -> RestMicrolitePage.from(null, source -> "test")).isExactlyInstanceOf(MissingMandatoryValueException.class);
  }

  @Test
  void shouldNotConvertWithoutMappingFunction() {
    assertThatThrownBy(() -> RestMicrolitePage.from(page(), null)).isExactlyInstanceOf(MissingMandatoryValueException.class);
  }

  @Test
  void shouldMapFromDomainPage() {
    RestMicrolitePage<String> page = RestMicrolitePage.from(page(), Function.identity());

    assertThat(page.getContent()).containsExactly("test");
    assertThat(page.getCurrentPage()).isEqualTo(2);
    assertThat(page.getPageSize()).isEqualTo(10);
    assertThat(page.getTotalElementsCount()).isEqualTo(21);
    assertThat(page.getPagesCount()).isEqualTo(3);
  }

  @Test
  void shouldGetPageCountForPageLimit() {
    RestMicrolitePage<String> page = RestMicrolitePage.from(pageBuilder().totalElementsCount(3).pageSize(3).build(), Function.identity());

    assertThat(page.getPagesCount()).isEqualTo(1);
  }

  @Test
  void shouldSerializeToJson() {
    assertThat(JsonHelper.writeAsString(RestMicrolitePage.from(page(), Function.identity()))).isEqualTo(json());
  }

  private String json() {
    return """
        {"content":["test"],\
        "currentPage":2,\
        "pageSize":10,\
        "totalElementsCount":21,\
        "pagesCount":3,\
        "hasPrevious":true,\
        "hasNext":false\
        }\
        """;
  }
}
